backend/requirements.txt
Django==3.1.4
psycopg2==2.8.6
djangorestframework==3.12.2
django-rest-auth==0.9.5
whitenoise==4.1.4
dj-database-url==0.5.0